# docs/api/openapi.yaml
openapi: 3.1.0
info:
  title: API de Pagamentos Sindireceita
  version: "1.0.0"
  summary: Cobrança e conciliação (PIX, boleto, checkout, DDA) sobre bancos do cliente.
  description: |
    API REST `/v1` que empresas usam como meio de cobrança sobre a **sua própria**
    credencial bancária (modelo PSP-Indireto — o dinheiro liquida direto na conta
    da empresa). Primeiro banco em produção: **C6**.

    Guia de integração/onboarding: `integration-guide.md` (mesma pasta).

    Convenções globais:
    - **Autenticação:** `Authorization: Bearer <token>` em toda rota `/v1`. A
      empresa-cliente é derivada do token, nunca de um parâmetro. Token
      ausente/inválido → `401`.
    - **Valores monetários:** sempre em **centavos** (inteiro). R$ 10,00 → `1000`.
    - **Datas/hora:** RFC 3339 (ex. `2026-08-20T00:00:00-03:00`), exceto a janela
      do extrato que usa `YYYY-MM-DD`.
    - **Idempotência:** rotas de escrita exigem o header `Idempotency-Key`.
    - **Erros:** envelope uniforme `{ "error": "..." }` (não é RFC 7807; ver guia).
    - **Limites de entrada:** corpo até 1 MiB (`413` acima); campos JSON
      desconhecidos são rejeitados (`400`).
    - **Rate limit:** token-bucket por empresa (burst 20, 10 req/s) → `429`.
  contact:
    name: Suporte técnico Sindireceita
  license:
    name: Proprietária
servers:
  - url: https://api.pagamentos.sindireceita.example
    description: Produção (URL real definida por ambiente)
...
    description: Homologação
security:
  - bearerAuth: []
...
paths:
...
  /v1/boletos/{id}:
    get:
      operationId: getBoleto
      tags: [Boleto]
      summary: Consultar boleto
      parameters:
        - $ref: "#/components/parameters/PathId"
      responses:
        "200":
          description: Boleto.
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Boleto" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "404": { $ref: "#/components/responses/NotFound" }
        "500": { $ref: "#/components/responses/InternalError" }
    put:
      operationId: updateBoleto
      tags: [Boleto]
      summary: Alterar boleto (vencimento/validade/valor/multa/juros/desconto)
      parameters:
        - $ref: "#/components/parameters/PathId"
        - $ref: "#/components/parameters/IdempotencyKey"
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/UpdateBoletoRequest" }
      responses:
        "200":
          description: Boleto alterado.
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Boleto" }
        "400": { $ref: "#/components/responses/BadRequest" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "404": { $ref: "#/components/responses/NotFound" }
        "409": { $ref: "#/components/responses/Conflict" }
        "500": { $ref: "#/components/responses/InternalError" }
    delete:
      operationId: deleteBoleto
      tags: [Boleto]
      summary: Baixa / cancelamento de boleto
      parameters:
        - $ref: "#/components/parameters/PathId"
      responses:
        "204":
          description: Boleto baixado/cancelado.
        "401": { $ref: "#/components/responses/Unauthorized" }
        "404": { $ref: "#/components/responses/NotFound" }
        "409": { $ref: "#/components/responses/Conflict" }
        "500": { $ref: "#/components/responses/InternalError" }
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      description: Token de API escopado por empresa-cliente.
  parameters:
    IdempotencyKey:
      name: Idempotency-Key
      in: header
      required: true
      description: Chave estável por intenção de negócio. Reenvio com a mesma chave não duplica o efeito.
      schema: { type: string, minLength: 1 }
    PathId:
      name: id
      in: path
      required: true
      schema: { type: string }
...
  responses:
    BadRequest:
      description: Validação falhou, JSON malformado, campo desconhecido ou header obrigatório ausente.
      content:
        application/json:
          schema: { $ref: "#/components/schemas/Error" }
          examples:
            invalid: { value: { error: "invalid request" } }
            idem: { value: { error: "missing Idempotency-Key header" } }
    Unauthorized:
      description: Token ausente ou inválido.
      content:
        application/json:
          schema: { $ref: "#/components/schemas/Error" }
          example: { error: "unauthorized" }
    NotFound:
      description: Recurso inexistente ou pertencente a outra empresa (mesma resposta, para não vazar existência).
      content:
        application/json:
          schema: { $ref: "#/components/schemas/Error" }
          example: { error: "not found" }
    Conflict:
      description: Conflito de estado.
      content:
        application/json:
          schema: { $ref: "#/components/schemas/Error" }
          example: { error: "conflict" }
...
    InternalError:
      description: Erro interno (transitório).
      content:
        application/json:
          schema: { $ref: "#/components/schemas/Error" }
          example: { error: "internal error" }
  schemas:
    Error:
      type: object
      description: Envelope de erro uniforme (não é RFC 7807).
      required: [error]
      properties:
        error: { type: string, description: Mensagem curta e segura (sem detalhe interno). }
...
    DiscountTier:
      type: object
      description: Faixa de desconto por antecipação; informe exatamente um de bps ou fixed_cents.
      properties:
        days_before_due: { type: integer }
        bps: { type: integer, format: int64 }
        fixed_cents: { type: integer, format: int64 }
...
    UpdateBoletoRequest:
      type: object
      required: [due_date]
      description: Alteração de vencimento/validade/valor/multa/juros/desconto (grupo 5).
      properties:
        amount_cents: { type: integer, format: int64 }
        currency: { type: string, example: BRL }
        due_date: { type: string, format: date-time }
        valid_until: { type: string, format: date-time }
        fine_bps: { type: integer, format: int64 }
        fine_fixed_cents: { type: integer, format: int64 }
        monthly_interest_bps: { type: integer, format: int64 }
        discounts:
          type: array
          items: { $ref: "#/components/schemas/DiscountTier" }
    Boleto:
      type: object
      properties:
        boleto_id: { type: string }
        txid: { type: string }
        status: { type: string }
        qr_code: { type: string, description: Payload BolePix EMV copia-e-cola. }
        barcode: { type: string, description: Linha digitável. }
        our_number: { type: string, description: Nosso número. }
        digitable_line: { type: string }
        amount_cents: { type: integer, format: int64 }
        due_date: { type: string, format: date-time }
        valid_until: { type: string, format: date-time }
        fine_bps: { type: integer, format: int64 }
        fine_fixed_cents: { type: integer, format: int64 }
        monthly_interest_bps: { type: integer, format: int64 }
        discounts:
          type: array
          items: { $ref: "#/components/schemas/DiscountTier" }
...